+++ src/Exception/DeclinationIsNotActive.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * @copyright Copyright (c) Wizacha
+ * @license Proprietary
+ */
+declare(strict_types=1);
+
+namespace Wizaplace\SDK\Exception;
+
+final class DeclinationIsNotActive extends \Exception implements DomainError
+{
+    /**
+     * @var array
+     */
+    private $context;
+
+    /**
+     * @internal
+     */
+    public function __construct(string $message, array $context = [], ?\Throwable $previous = null)
+    {
+        $this->context = $context;
+
+        parent::__construct($message, $previous);
+    }
+
+    final public function getContext(): array
+    {
+        return $this->context;
+    }
+
+    public static function getErrorCode(): ErrorCode
+    {
+        return ErrorCode::DECLINATION_IS_NOT_ACTIVE();
+    }
+}